Rename old GL to RoE
