+++ src/parser/jobs/mnk/modules/RiddleOfEarth.tsx
